Finishes debugging classes
When operator() was replaced with set/getCell all statements
which relied on it the operator to return a cell broke.
Currently using gross tempCell object to store the return from
get, then using the temp for the functions. If data modified the
temp cell is used with set to overwrite old cell. This doesn't seem
efficient or clean, is there another way?

Added setCellLive method to simplify setting cells alive due to ^